requirement.txt
gradio
requests
Pillow
fastapi
uvicorn
numpy
pandas
scipy
torch
torchvision


# yolov8
ultralytics

